requirements.txt
# 🚀 BASED CODER CLI - Enhanced Requirements with Claude 4 Coder Agent
# Made by @Lucariolucario55 on Telegram

# Core dependencies
asyncio>=3.4.3
aiohttp>=3.8.0
aiofiles>=0.8.0
colorama>=0.4.6
rich>=13.0.0
numpy>=1.21.0
psutil>=5.9.0
python-dotenv>=1.0.0

# AI and ML dependencies
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Database and storage
qdrant-client>=1.6.0
aiosqlite>=0.19.0

# Web and networking
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
duckduckgo-search>=4.0.0
selenium>=4.10.0
playwright>=1.40.0

# Code analysis and processing
ast>=0.0.2
pylint>=2.17.0
black>=23.0.0
flake8>=6.0.0

# HuggingFace and model management
huggingface-hub>=0.16.0
accelerate>=0.20.0

# Configuration and data handling
pyyaml>=6.0
pydantic>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Enhanced features for Claude 4 Coder Agent
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.10

# MCP (Model Context Protocol) - Local setup
# Note: MCP tools will be installed via npm instead of pip

# Advanced debugging and profiling
memory-profiler>=0.61.0
line-profiler>=4.1.0

# Performance monitoring
prometheus-client>=0.17.0

# Enhanced logging
structlog>=23.1.0

# Optional: GPU support (uncomment if needed)
# torch-cuda>=2.0.0 